+++ AsyncLoggers/src/Patches/BepInExChainloaderPatch.cs
@@ -1,14 +1,14 @@
 ﻿using System;
-using System.IO;
 using System.Linq;
 using System.Reflection;
 using AsyncLoggers.BepInExListeners;
+using AsyncLoggers.DBAPI;
 using AsyncLoggers.Wrappers.BepInEx;
-using BepInEx;
 using BepInEx.Bootstrap;
 using BepInEx.Logging;
 using HarmonyLib;
 using UnityEngine;
+using Logger = BepInEx.Logging.Logger;
 
 namespace AsyncLoggers.Patches
 {
@@ -21,14 +21,11 @@ namespace AsyncLoggers.Patches
         {
             try
             {
-                if (PluginConfig.DbLogger.Enabled.Value && SqliteChecker.isLoaded())
+                SqliteLoggerImpl.WriteEvent("BepInEx","Chainloader.Initialize", "Initializing Chainloader");
+                if (SqliteLoggerImpl.Enabled)
                 {
                     AsyncLoggerPreloader.Log.LogWarning($"Adding Sqlite to BepInEx Listeners");
-                    BepInEx.Logging.Logger.Listeners.Add(
-                        new AsyncLogListenerWrapper(
-                        new SqliteListener(Path.Combine(Paths.BepInExRootPath, "LogOutput.sqlite"))));
-                    //BepInEx.Logging.Logger.Listeners.Add(
-                    //new AsyncLogListenerWrapper(new TsvListener(Path.Combine(Paths.BepInExRootPath, "LogOutput.tsv"))));
+                    BepInEx.Logging.Logger.Listeners.Add(new AsyncLogListenerWrapper(new SqliteListener()));
                 }
             }
             catch (Exception ex)
@@ -41,74 +38,86 @@ namespace AsyncLoggers.Patches
         [HarmonyPatch(typeof(Chainloader), nameof(Chainloader.Initialize))]
         private static void UseAsyncLogListeners()
         {
-            AsyncLoggerPreloader.Log.LogInfo(
-                "using logMessageReceivedThreaded instead of logMessageReceived for UnityLogSource!!");
-            Application.LogCallback oldHandler = UnityLogSource.OnUnityLogMessageReceived;
-            Application.LogCallback newHandler = UnityLogMessageReceivedWrapper;
-            EventInfo eventInfo =
-                typeof(Application).GetEvent("logMessageReceived", BindingFlags.Static | BindingFlags.Public);
-            if (eventInfo != null)
-                eventInfo.RemoveEventHandler(null, oldHandler);
-
-            eventInfo = typeof(Application).GetEvent("logMessageReceivedThreaded",
-                BindingFlags.Static | BindingFlags.Public);
-            if (eventInfo != null)
-                eventInfo.AddEventHandler(null, newHandler);
-
-            if (PluginConfig.BepInEx.Enabled.Value)
+            try
             {
-                AsyncLoggerPreloader.Log.LogWarning("Converting BepInEx loggers to async!!");
-                var collection = BepInEx.Logging.Logger.Listeners;
-                var list = collection.ToList();
-                foreach (var logger in list)
+                SqliteLoggerImpl.WriteEvent("BepInEx", "Chainloader.Initialize", "Finished Initializing Chainloader");
+                
+                AsyncLoggerPreloader.Log.LogInfo(
+                    "using logMessageReceivedThreaded instead of logMessageReceived for UnityLogSource!!");
+                Application.LogCallback handler = UnityLogSource.OnUnityLogMessageReceived;
+                EventInfo eventInfo =
+                    typeof(Application).GetEvent("logMessageReceived", BindingFlags.Static | BindingFlags.Public);
+                if (eventInfo != null)
+                    eventInfo.RemoveEventHandler(null, handler);
+
+                eventInfo = typeof(Application).GetEvent("logMessageReceivedThreaded",
+                    BindingFlags.Static | BindingFlags.Public);
+                if (eventInfo != null)
+                    eventInfo.AddEventHandler(null, handler);
+
+                if (PluginConfig.BepInEx.Enabled.Value)
                 {
-                    var isConsole = logger is ConsoleLogListener;
-                    var isUnity = logger is UnityLogListener;
-                    var isDisk = logger is DiskLogListener;
-
-                    if (!isConsole && !isDisk && !isUnity)
-                        continue;
-                    if (isConsole && !PluginConfig.BepInEx.Console.Value)
-                        continue;
-                    if (isUnity && !PluginConfig.BepInEx.Unity.Value)
-                        continue;
-                    if (isDisk && !PluginConfig.BepInEx.Disk.Value)
-                        continue;
-
-                    AsyncLoggerPreloader.Log.LogWarning($"{logger.GetType().Name} Converted");
-                    collection.Remove(logger);
-                    collection.Add(new AsyncLogListenerWrapper(logger));
-                }
-            }
-
-            void UnityLogMessageReceivedWrapper(string message, string stackTrace, LogType type)
-            {
-                if (type == LogType.Exception)
-                {
-                    var sections = stackTrace.Trim().Split('\n').ToList();
-                    var isJob = sections[sections.Count - 1].StartsWith("Unity.Jobs.JobStruct");
-                    var isThread = sections[sections.Count - 1].StartsWith("System.Threading.ThreadHelper");
-                    if (isJob || isThread)
+                    AsyncLoggerPreloader.Log.LogWarning("Converting BepInEx loggers to async!!");
+                    var collection = BepInEx.Logging.Logger.Listeners;
+                    var list = collection.ToList();
+                    foreach (var logger in list)
                     {
-                        var to_remove = 1;
-                        for (var i = sections.Count - 2; i >= 0; i--)
-                        {
-                            var curr = sections[i];
-                            if (!curr.StartsWith("AsyncLoggers.Wrappers"))
-                                break;
-                            to_remove++;
-                        }
+                        var isConsole = logger is ConsoleLogListener;
+                        var isUnity = logger is UnityLogListener;
+                        var isDisk = logger is DiskLogListener;
 
-                        if (to_remove > 1)
-                        {
-                            if (to_remove < sections.Count - 2)
-                                sections.RemoveRange(sections.Count - to_remove, to_remove);
-                            stackTrace = string.Join("\n", sections) + "\n";
-                        }
+                        if (!isConsole && !isDisk && !isUnity)
+                            continue;
+                        if (isConsole && !PluginConfig.BepInEx.Console.Value)
+                            continue;
+                        if (isUnity && !PluginConfig.BepInEx.Unity.Value)
+                            continue;
+                        if (isDisk && !PluginConfig.BepInEx.Disk.Value)
+                            continue;
+
+                        AsyncLoggerPreloader.Log.LogWarning($"{logger.GetType().Name} Converted");
+                        collection.Remove(logger);
+                        collection.Add(new AsyncLogListenerWrapper(logger));
                     }
                 }
-
-                UnityLogSource.OnUnityLogMessageReceived(message, stackTrace, type);
+            }
+            catch (Exception ex)
+            {
+                AsyncLoggerPreloader.Log.LogError($"Exception while converting BepInEx loggers to async! {ex}");
+            }
+        }
+        
+        
+        [HarmonyPrefix]
+        [HarmonyPatch(typeof(Logger), nameof(Logger.LogMessage))]
+        private static void OnStart(object data)
+        {
+            try
+            {   
+                if (data is"Chainloader started")
+                    SqliteLoggerImpl.WriteEvent("BepInEx", "Chainloader.Start", "Starting Chainloader");
+            }
+            catch (Exception ex)
+            {
+                AsyncLoggerPreloader.Log.LogError($"Exception logging event {ex}");
+            }
+        }
+        
+        [HarmonyPrefix]
+        [HarmonyPatch(typeof(Logger), nameof(Logger.LogMessage))]
+        private static void TrackLoadedPlugins(object data)
+        {
+            try
+            {
+                if (data is "Chainloader startup complete")
+                {
+                    SqliteLoggerImpl.WriteEvent("BepInEx", "Chainloader.Start", "Finished Starting Chainloader");
+                    SqliteLoggerImpl.WriteMods(Chainloader.PluginInfos.Values);
+                }
+            }
+            catch (Exception ex)
+            {
+                AsyncLoggerPreloader.Log.LogError($"Exception logging mods {ex}");
             }
         }
     }