+++ AsyncLoggers/src/Wrappers/IWrapper.cs
@@ -7,5 +7,11 @@
         public void Schedule(LogCallback callback);
 
         public void Stop(bool immediate = false);
+        
+        public enum ContextType
+        {
+            Log,
+            Event
+        }
     }
 }