+++ AsyncLoggers/src/Wrappers/JobWrapper.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Threading;
+using AsyncLoggers.StaticContexts;
 using DisruptorUnity3d;
 using Unity.Jobs;
 using UnityEngine;
@@ -21,25 +21,37 @@ namespace AsyncLoggers.Wrappers
         private readonly LogJob _loggingJobStruct;
         private readonly ConcurrentCircularBuffer<IWrapper.LogCallback> _taskRingBuffer;
         private volatile RunCondition _shouldRun;
+        internal static IWrapper.ContextType ContextType { get; private set; }
 
-        public JobWrapper()
+        public JobWrapper(IWrapper.ContextType contextType = IWrapper.ContextType.Log)
         {
             var _id = Interlocked.Add(ref _IDSeed, 1);
             INSTANCES[_id] = this;
+            ContextType = contextType;
             _taskRingBuffer = new ConcurrentCircularBuffer<IWrapper.LogCallback>(PluginConfig.Scheduler.JobBufferSize.Value);
             DefaultCondition = ()=>_taskRingBuffer.Count > 0;
             _shouldRun = DefaultCondition;
             _loggingJobStruct = new LogJob(_id);
         }
 
+        
+        private long getUUID()
+        {
+            return ContextType switch
+            {
+                IWrapper.ContextType.Log => LogContext.Uuid!.Value,
+                IWrapper.ContextType.Event => EventContext.Uuid!.Value,
+                _ => -1L
+            };
+        }
+        
         public void Schedule(IWrapper.LogCallback callback)
         {
             if (_shouldRun != DefaultCondition) 
                 return;
             
-            
-            var logUUID = LogContext.Uuid;
-            var timestamp = LogContext.Timestamp;
+            var uuid = getUUID();
+            var timestamp = GenericContext.Timestamp;
 
             _taskRingBuffer.Enqueue(CallbackWrapper);
 
@@ -52,16 +64,25 @@ namespace AsyncLoggers.Wrappers
             {
                 try
                 {
-                    LogContext.Async = true;
-                    LogContext.Timestamp = timestamp;
-                    LogContext.Uuid = logUUID;
+                    GenericContext.Async = true;
+                    GenericContext.Timestamp = timestamp;
+                    switch (ContextType)
+                    {
+                        case IWrapper.ContextType.Log:
+                            LogContext.Uuid = uuid;
+                            break;
+                        case IWrapper.ContextType.Event:
+                            EventContext.Uuid = uuid;
+                            break;
+                    }
                     callback();
                 }
                 finally
                 {
-                    LogContext.Async = false;
-                    LogContext.Timestamp = null;
+                    GenericContext.Async = false;
+                    GenericContext.Timestamp = null;
                     LogContext.Uuid = null;
+                    EventContext.Uuid = null;
                 }
             }
         }